backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
Babel==2.14.0
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.1
docxcompose==1.4.0
docxtpl==0.16.8
et-xmlfile==1.1.0
fastapi==0.110.1
Flask==3.0.3
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
reportlab==4.1.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
Werkzeug==3.0.2
